requirements.txt
asgiref==3.8.1
attrs==23.2.0
diff-match-patch==20230430
Django==5.0.4
django-extensions==3.2.3
django-import-export==4.0.3
djangorestframework==3.15.1
inflection==0.5.1
jsonschema==4.22.0
jsonschema-specifications==2023.12.1
packaging==24.0
pillow==10.3.0
psycopg==3.1.18
psycopg2==2.9.9
python-dotenv==1.0.1
pytz==2024.1
PyYAML==6.0.1
referencing==0.35.1
rpds-py==0.18.1
sqlparse==0.5.0
swagger-spec-validator==3.0.3
tablib==3.5.0
typing_extensions==4.11.0
tzdata==2024.1
uritemplate==4.1.1
annotated-types==0.7.0
anyio==4.4.0
asgiref==3.8.1
certifi==2024.6.2
click==8.1.7
colorama==0.4.6
defusedxml==0.7.1
diff-match-patch==20230430
Django==5.0.3
django-extensions==3.2.3
django-import-export==3.3.7
dnspython==2.6.1
email_validator==2.1.2
et-xmlfile==1.1.0
fastapi==0.111.0
fastapi-cli==0.0.4
h11==0.14.0
httpcore==1.0.5
httptools==0.6.1
httpx==0.27.0
idna==3.7
Jinja2==3.1.4
markdown-it-py==3.0.0
MarkupPy==1.14
MarkupSafe==2.1.5
mdurl==0.1.2
odfpy==1.4.1
openpyxl==3.1.2
orjson==3.10.5
pillow==10.3.0
psycopg2==2.9.9
pydantic==2.7.4
pydantic_core==2.18.4
Pygments==2.18.0
python-dotenv==1.0.1
python-multipart==0.0.9
PyYAML==6.0.1
rich==13.7.1
shellingham==1.5.4
sniffio==1.3.1
sqlparse==0.4.4
starlette==0.37.2
tablib==3.5.0
typer==0.12.3
typing_extensions==4.12.2
tzdata==2024.1
ujson==5.10.0
uvicorn==0.30.1
watchfiles==0.22.0
websockets==12.0
xlrd==2.0.1
xlwt==1.3.0
aiogram~=2.14
environs~=8.0.0

